refactor: getCarrotReward to hook
